Rename the *test* directory to *black_box_tests*
