+++ .travis.yml
@@ -13,12 +13,12 @@ python:
   - "pypy3"
 env:
   - TEST_ENV=lollygag
-  - TEST_ENV=test
+  - TEST_ENV=black_box_tests
 install:
 - pip install -r requirements.txt
 before_script:
-  - if [ "$TEST_ENV" = "test" ]; then sudo python setup.py install; fi
-  - if [ "$TEST_ENV" = "test" ]; then sudo chmod -R +xr .; fi
+  - if [ "$TEST_ENV" = "black_box_tests" ]; then sudo python setup.py install; fi
+  - if [ "$TEST_ENV" = "black_box_tests" ]; then sudo chmod -R +xr .; fi
 script:
 - cd $TEST_ENV
 - pytest
