+++ black_box_tests/mapper_example.py
